Implemented Vulnerable and Weak in the current branch
